+++ fuel-merkle/src/sparse.rs
@@ -9,8 +9,9 @@ pub(crate) use node::{StorageNode, StorageNodeError};
 pub use merkle_tree::{MerkleTree, MerkleTreeError};
 pub use node::Node;
 pub use primitive::Primitive;
+pub mod branch;
 pub mod in_memory;
-pub mod test;
+mod test;
 
 use crate::common::Bytes32;
