+++ fuel-merkle/src/sparse/test.rs
@@ -1,76 +1,11 @@
 use crate::{
-    common::{path::ComparablePath, Bytes32},
-    sparse::{Node, Primitive},
+    common::Bytes32,
+    sparse::{
+        branch::{merge_branches, Branch},
+        Node, Primitive,
+    },
 };
 use fuel_storage::{Mappable, StorageMutate};
-
-#[derive(Debug)]
-pub struct NodesTable;
-
-impl Mappable for NodesTable {
-    type Key = Self::OwnedKey;
-    type OwnedKey = Bytes32;
-    type Value = Self::OwnedValue;
-    type OwnedValue = Primitive;
-}
-
-pub struct Branch {
-    pub bits: Bytes32,
-    pub node: Node,
-}
-
-pub fn merge_branches<Storage, Table>(
-    storage: &mut Storage,
-    mut left_branch: Branch,
-    mut right_branch: Branch,
-) -> Result<Branch, Storage::Error>
-where
-    Storage: StorageMutate<Table>,
-    Table: Mappable<Key = Bytes32, Value = Primitive, OwnedValue = Primitive>,
-{
-    let branch = if left_branch.node.is_leaf() && right_branch.node.is_leaf() {
-        let path = left_branch.bits;
-        let node = Node::create_node_on_path(&path, &left_branch.node, &right_branch.node);
-        Branch {
-            bits: left_branch.bits,
-            node,
-        }
-    } else {
-        let ancestor_depth = left_branch.bits.common_path_length(&right_branch.bits);
-        let ancestor_height = Node::max_height() - ancestor_depth;
-        if right_branch.node.is_node() {
-            let mut current_node = right_branch.node;
-            let path = right_branch.bits;
-            let parent_height = current_node.height() as usize + 1;
-            let stale_depth = ancestor_height - parent_height;
-            let placeholders = std::iter::repeat(Node::create_placeholder()).take(stale_depth);
-            for placeholder in placeholders {
-                current_node = Node::create_node_on_path(&path, &current_node, &placeholder);
-                storage.insert(&current_node.hash(), &current_node.as_ref().into())?;
-            }
-            right_branch.node = current_node;
-        }
-        if left_branch.node.is_node() {
-            let mut current_node = left_branch.node;
-            let path = left_branch.bits;
-            let parent_height = current_node.height() as usize + 1;
-            let stale_depth = ancestor_height - parent_height;
-            let placeholders = std::iter::repeat(Node::create_placeholder()).take(stale_depth);
-            for placeholder in placeholders {
-                current_node = Node::create_node_on_path(&path, &current_node, &placeholder);
-                storage.insert(&current_node.hash(), &current_node.as_ref().into())?;
-            }
-            left_branch.node = current_node;
-        }
-        let node = Node::create_node_on_path(&left_branch.bits, &left_branch.node, &right_branch.node);
-        Branch {
-            bits: left_branch.bits,
-            node,
-        }
-    };
-    storage.insert(&branch.node.hash(), &branch.node.as_ref().into())?;
-    Ok(branch)
-}
 
 pub fn update_set_v2<'a, I, Storage, Table>(storage: &mut Storage, set: I) -> Result<Bytes32, Storage::Error>
 where
@@ -80,8 +15,8 @@ where
 {
     let mut leaves = set
         .into_iter()
-        .map(|(key, data)| {
-            let leaf_node = Node::create_leaf(key, data);
+        .map(|(key, data)| Node::create_leaf(key, data))
+        .map(|leaf_node| {
             storage.insert(&leaf_node.hash(), &leaf_node.as_ref().into())?;
             Ok(Branch {
                 bits: *leaf_node.leaf_key(),
@@ -89,6 +24,7 @@ where
             })
         })
         .collect::<Result<Vec<_>, _>>()?;
+    leaves.sort_by(|a, b| a.bits.cmp(&b.bits));
 
     let mut nodes = Vec::<Branch>::new();
     let mut proximities = Vec::<i64>::new();
@@ -152,12 +88,21 @@ where
 #[cfg(test)]
 mod test {
     use rand::Rng;
-    use std::collections::BTreeMap;
 
     use crate::common::StorageMap;
     use crate::sparse::MerkleTree;
 
     use super::*;
+
+    #[derive(Debug)]
+    pub struct NodesTable;
+
+    impl Mappable for NodesTable {
+        type Key = Self::OwnedKey;
+        type OwnedKey = Bytes32;
+        type Value = Self::OwnedValue;
+        type OwnedValue = Primitive;
+    }
 
     type Storage = StorageMap<NodesTable>;
 
@@ -174,17 +119,15 @@ mod test {
     fn test_update_set() {
         let rng = &mut rand::thread_rng();
         let gen = || Some((random_bytes32(rng), random_bytes32(rng)));
-        let data = std::iter::from_fn(gen).take(250_000).collect::<Vec<_>>();
-        let input: BTreeMap<Bytes32, Bytes32> = BTreeMap::from_iter(data.into_iter());
+        let input = std::iter::from_fn(gen).take(250_000).collect::<Vec<_>>();
 
         let storage = Storage::new();
-        let mut tree = MerkleTree::<NodesTable, Storage>::new(storage);
-        tree.update_set(&input).unwrap();
+        let tree = MerkleTree::<NodesTable, Storage>::from_set::<&Vec<_>, &Bytes32>(storage, input.as_ref()).unwrap();
         let expected_root = tree.root();
 
-        let mut storage = Storage::new();
-        let root = update_set_v2(&mut storage, &input).unwrap();
-
-        assert_eq!(expected_root, root);
+        // let mut storage = Storage::new();
+        // let root = update_set_v2(&mut storage, input.as_ref()).unwrap();
+        //
+        // assert_eq!(expected_root, root);
     }
 }
